+++ src/services/auth/oauth2/types.ts
@@ -4,10 +4,18 @@ import {
   OptionsWithRql,
   PagedResult,
 } from '../../types';
+import { AuthOauth2TokenService } from './accessTokens/types';
+import { OAuth2AuthorizationsService } from './authorizations/types';
+import { OAuth2RefreshTokenService } from './refreshTokens/types';
+
+export * from './accessTokens/types';
+export * from './authorizations/types';
+export * from './refreshTokens/types';
 
 export interface AuthOauth2Service {
   tokens: AuthOauth2TokenService;
   refreshTokens: OAuth2RefreshTokenService;
+  authorizations: OAuth2AuthorizationsService;
 
   /**
    * Create an OAuth2 authorization
@@ -19,6 +27,7 @@ export interface AuthOauth2Service {
    * @throws {ApplicationUnknownError}
    * @throws {CallbackNotValidError}
    * @throws {UnsupportedResponseTypeError}
+   * @deprecated - Will be removed in a future version, please use `auth.oauth2.authorizations.create` instead
    */
   createAuthorization(
     data: OAuth2AuthorizationCreation,
@@ -33,6 +42,7 @@ export interface AuthOauth2Service {
    * none | | Can only see a list of OAuth2 authorizations for this account
    * VIEW_AUTHORIZATIONS | global | See any authorizations belonging to any user
    * @see https://swagger.extrahorizon.com/swagger-ui/?url=https://swagger.extrahorizon.com/auth-service/2.0.4-dev/openapi.yaml#/OAuth2/get_oauth2_authorizations
+   * @deprecated - Will be removed in a future version, please use `auth.oauth2.authorizations.find*` instead
    */
   getAuthorizations(
     options?: OptionsWithRql
@@ -47,6 +57,7 @@ export interface AuthOauth2Service {
    * DELETE_AUTHORIZATIONS | global | Delete any authorizations belonging to any user
    * @see https://swagger.extrahorizon.com/swagger-ui/?url=https://swagger.extrahorizon.com/auth-service/2.0.4-dev/openapi.yaml#/OAuth2/delete_oauth2_authorizations__authorizationId_
    * @throws {ResourceUnknownError}
+   * @deprecated - Will be removed in a future version, please use `auth.oauth2.authorizations.remove` instead
    */
   deleteAuthorization(
     authorizationId: string,
@@ -54,132 +65,12 @@ export interface AuthOauth2Service {
   ): Promise<AffectedRecords>;
 }
 
-export interface AuthOauth2TokenService {
-  /**
-   * Get a list of OAuth2 tokens
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
-   */
-  find(options?: OptionsWithRql): Promise<PagedResult<OAuth2Token>>;
-
-  /**
-   * Get a list of OAuth2 tokens
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
-   */
-  findAll(options?: OptionsWithRql): Promise<OAuth2Token[]>;
-
-  /**
-   * Get the first OAuth2 token found
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
-   */
-  findFirst(options?: OptionsWithRql): Promise<OAuth2Token | undefined>;
-
-  /**
-   * Get an oAuth2 token by its id
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 tokens for this account
-   * VIEW_AUTHORIZATIONS | global | See any OAuth2 tokens belonging to any user
-   */
-  findById(id: string, options?: OptionsWithRql): Promise<OAuth2Token | undefined>;
-
-  /**
-   * Remove an oAuth2 token
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only delete OAuth2 tokens for this account
-   * DELETE_AUTHORIZATIONS | global | Delete any OAuth2 tokens belonging to any user
-   */
-  remove(id: string): Promise<AffectedRecords>;
-}
-
-export interface OAuth2RefreshTokenService {
-  /**
-   * # Get a list of OAuth2 refresh tokens
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 refresh tokens for this account
-   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
-   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
-   * @param options.rql Add filters to the requested list
-   * @returns PagedResult<OAuth2RefreshToken>
-   */
-  find(options?: OptionsWithRql): Promise<PagedResult<OAuth2RefreshToken>>;
-
-  /**
-   * Get a list of OAuth2 refresh tokens
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 refresh tokens for this account
-   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
-   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
-   * @param options.rql Add filters to the requested list
-   * @returns OAuth2RefreshToken[]
-   */
-  findAll(options?: OptionsWithRql): Promise<OAuth2RefreshToken[]>;
-
-  /**
-   * Get the first OAuth2 refresh token found
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 refresh tokens for this account
-   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
-   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
-   * @param options.rql Add filters to the requested list
-   * @returns {Promise<OAuth2RefreshToken | undefined>}
-   */
-  findFirst(options?: OptionsWithRql): Promise<OAuth2RefreshToken | undefined>;
-
-  /**
-   * Get an oAuth2 refresh token by its id
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only see a list of OAuth2 refresh tokens for this account
-   * VIEW_OAUTH2_REFRESH_TOKENS or VIEW_AUTHORIZATIONS | global | Can see a list of OAuth2 refresh tokens for any account
-   * Using VIEW_AUTHORIZATIONS for this endpoint is deprecated; use VIEW_OAUTH2_REFRESH_TOKENS instead
-   * @param id the refresh token id
-   * @param options.rql Add filters to the requested list
-   * @returns {Promise<OAuth2RefreshToken | undefined>}
-   */
-  findById(id: string, options?: OptionsWithRql): Promise<OAuth2RefreshToken | undefined>;
-
-  /**
-   * Delete an oAuth2 refresh token
-   *
-   * Permission | Scope | Effect
-   * - | - | -
-   * none | | Can only delete OAuth2 refresh tokens for this account
-   * DELETE_OAUTH2_REFRESH_TOKEN or DELETE_AUTHORIZATIONS | global | Delete any OAuth2 refresh tokens belonging to any user
-   * Using DELETE_AUTHORIZATIONS for this endpoint is deprecated; use DELETE_OAUTH2_REFRESH_TOKEN instead
-   * @param id the refresh token id
-   * @returns AffectedRecords
-   */
-  remove(id: string, options?: OptionsBase): Promise<AffectedRecords>;
-}
-
 export interface OAuth2AuthorizationCreation {
-  responseType: string;
+  responseType: 'code';
   clientId: string;
   redirectUri?: string;
   state?: string;
-  codeChallengeMethod?: string;
+  codeChallengeMethod?: PKCECodeMethods;
   codeChallenge?: string;
 }
 
@@ -189,7 +80,7 @@ export interface OAuth2AuthorizationCreationResponse {
   userId: string;
   redirectUri: string;
   state?: string;
-  codeChallengeMethod?: string;
+  codeChallengeMethod?: PKCECodeMethods;
   codeChallenge?: string;
   authorizationCode: string;
   expiryTimestamp: Date;
@@ -203,7 +94,7 @@ export interface OAuth2Authorization {
   userId: string;
   redirectUri: string;
   state?: string;
-  codeChallengeMethod?: string;
+  codeChallengeMethod?: PKCECodeMethods;
   /**
    @deprecated `codeChallenge` will be removed from responses returned by listing endpoints in a future version.
    */
@@ -217,25 +108,7 @@ export interface OAuth2Authorization {
   creationTimestamp: Date;
 }
 
-export interface OAuth2Token {
-  id: string;
-  applicationId: string;
-  userId: string;
-  refreshTokenId: string;
-  /**
-   * @deprecated `accessToken` will be removed from responses returned by listing endpoints in a future version.
-   */
-  accessToken?: string;
-  expiryTimestamp: Date;
-  updateTimestamp: Date;
-  creationTimestamp: Date;
-}
-
-export interface OAuth2RefreshToken {
-  id: string;
-  applicationId: string;
-  userId: string;
-  expiryTimestamp: Date;
-  updateTimestamp: Date;
-  creationTimestamp: Date;
+export enum PKCECodeMethods {
+  PLAIN = 'plain',
+  S256 = 'S256',
 }